fix: added slide order in Slide doctype, removed Settings doctype
