+++ frappe/core/setup_wizard_slide/welcome_to_erpnext!/welcome_to_erpnext!.json
@@ -1,0 +1,20 @@
+{
+ "add_more_button": 0,
+ "app": "Frappe",
+ "creation": "2019-11-26 16:59:48.315297",
+ "docstatus": 0,
+ "doctype": "Setup Wizard Slide",
+ "domains": [],
+ "help_links": [],
+ "idx": 0,
+ "max_count": 3,
+ "modified": "2019-11-26 17:00:49.350931",
+ "modified_by": "Administrator",
+ "name": "Welcome to ERPNext!",
+ "owner": "Administrator",
+ "slide_fields": [],
+ "slide_module": "Core",
+ "slide_order": 0,
+ "slide_title": "Welcome to ERPNext!",
+ "slide_type": "Information"
+}
