--- frappe/desk/doctype/setup_wizard_settings/setup_wizard_settings.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2019, Frappe Technologies and contributors
-// For license information, please see license.txt
-
-frappe.ui.form.on('Setup Wizard Settings', {
-	// refresh: function(frm) {
-
-	// }
-});
