--- frappe/desk/doctype/setup_wizard_settings/setup_wizard_settings.json
@@ -1,46 +1,0 @@
-{
- "creation": "2019-11-13 15:55:07.016295",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "slide_order_section",
-  "slide_order"
- ],
- "fields": [
-  {
-   "fieldname": "slide_order_section",
-   "fieldtype": "Section Break",
-   "label": "Slide Order"
-  },
-  {
-   "fieldname": "slide_order",
-   "fieldtype": "Table",
-   "label": "Slide Order",
-   "options": "Setup Wizard Slide Order",
-   "reqd": 1
-  }
- ],
- "issingle": 1,
- "modified": "2019-11-19 11:56:32.757433",
- "modified_by": "Administrator",
- "module": "Desk",
- "name": "Setup Wizard Settings",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "print": 1,
-   "read": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "quick_entry": 1,
- "sort_field": "modified",
- "sort_order": "DESC",
- "track_changes": 1
-}
