--- frappe/desk/doctype/setup_wizard_settings/test_setup_wizard_settings.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (c) 2019, Frappe Technologies and Contributors
-# See license.txt
-from __future__ import unicode_literals
-
-# import frappe
-import unittest
-
-class TestSetupWizardSettings(unittest.TestCase):
-	pass
