+++ frappe/desk/doctype/setup_wizard_slide/setup_wizard_slide.json
@@ -7,8 +7,10 @@
  "field_order": [
   "slide_title",
   "app",
+  "slide_order",
   "column_break_4",
   "image_src",
+  "slide_module",
   "description_section_break",
   "slide_desc",
   "action_section_break",
@@ -17,7 +19,7 @@
   "column_break_6",
   "max_count",
   "add_more_button",
-  "section_break_5",
+  "section_break_18",
   "ref_doctype",
   "slide_fields",
   "section_break_10",
@@ -35,7 +37,7 @@
    "unique": 1
   },
   {
-   "depends_on": "eval:doc.slide_type !== 'Information';",
+   "depends_on": "eval:doc.slide_type!='Information'",
    "description": "By default the code inside `create_onboarding_docs` method of the `Reference Document Type` is executed. If your method is not on the doctype level, place this method in {app_name}.utilities.onboarding_utils.{method_name} and specify the method name here",
    "fieldname": "submit_method",
    "fieldtype": "Data",
@@ -56,19 +58,13 @@
   },
   {
    "default": "0",
-   "depends_on": "eval: doc.slide_type !== 'Information';",
+   "depends_on": "eval:doc.slide_type!='Information'",
    "fieldname": "add_more_button",
    "fieldtype": "Check",
    "label": "Add More Button"
   },
   {
-   "depends_on": "eval: doc.slide_type !== 'Information'",
-   "fieldname": "section_break_5",
-   "fieldtype": "Section Break",
-   "label": "Fields"
-  },
-  {
-   "depends_on": "eval:doc.slide_type !== 'Information';",
+   "depends_on": "eval:doc.slide_type!='Information'",
    "fieldname": "slide_fields",
    "fieldtype": "Table",
    "label": "Slide Fields",
@@ -135,14 +131,33 @@
    "label": "Description"
   },
   {
-   "depends_on": "eval:doc.slide_type !== 'Information';",
+   "depends_on": "eval:doc.slide_type!='Information'",
    "fieldname": "ref_doctype",
    "fieldtype": "Link",
    "label": "Reference Document Type",
    "options": "DocType"
+  },
+  {
+   "default": "0",
+   "description": "Determines the order of the slide in the wizard. If the slide is not to be displayed, priority should be set to 0.",
+   "fieldname": "slide_order",
+   "fieldtype": "Int",
+   "label": "Slide Order"
+  },
+  {
+   "depends_on": "eval:doc.slide_type=='Information'",
+   "fieldname": "slide_module",
+   "fieldtype": "Link",
+   "label": "Module",
+   "options": "Module Def"
+  },
+  {
+   "fieldname": "section_break_18",
+   "fieldtype": "Section Break",
+   "label": "Fields"
   }
  ],
- "modified": "2019-11-26 13:52:32.351576",
+ "modified": "2019-11-26 17:33:34.553367",
  "modified_by": "Administrator",
  "module": "Desk",
  "name": "Setup Wizard Slide",
