--- frappe/desk/doctype/setup_wizard_slide_order/setup_wizard_slide_order.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (c) 2019, Frappe Technologies and contributors
-# For license information, please see license.txt
-
-from __future__ import unicode_literals
-# import frappe
-from frappe.model.document import Document
-
-class SetupWizardSlideOrder(Document):
-	pass
