--- frappe/desk/setup_wizard_settings/setup_wizard_settings/setup_wizard_settings.json
@@ -1,27 +1,0 @@
-{
- "creation": "2019-11-13 16:07:24.786989",
- "docstatus": 0,
- "doctype": "Setup Wizard Settings",
- "idx": "0",
- "modified": "2019-11-26 14:47:30.065799",
- "modified_by": "Administrator",
- "name": "Setup Wizard Settings",
- "owner": "Administrator",
- "slide_order": [
-  {
-   "slide": "Welcome to ERPNext!"
-  },
-  {
-   "slide": "Company Letter Head"
-  },
-  {
-   "slide": "Add A Few Products You Buy Or Sell"
-  },
-  {
-   "slide": "Add A Few Customers"
-  },
-  {
-   "slide": "Add A Few Suppliers"
-  }
- ]
-}
