--- frappe/desk/setup_wizard_slide/add_a_few_customers/add_a_few_customers.json
@@ -1,61 +1,0 @@
-{
- "add_more_button": 1,
- "app": "ERPNext",
- "creation": "2019-11-15 14:44:10.065014",
- "docstatus": 0,
- "doctype": "Setup Wizard Slide",
- "domains": [
-  {
-   "domain": "Manufacturing"
-  },
-  {
-   "domain": "Services"
-  },
-  {
-   "domain": "Retail"
-  },
-  {
-   "domain": "Distribution"
-  }
- ],
- "help_links": [
-  {
-   "label": "Customers",
-   "video_id": "zsrrVDk6VBs"
-  }
- ],
- "idx": 0,
- "image_src": "/assets/erpnext/images/illustrations/customer.png",
- "max_count": 3,
- "modified": "2019-11-26 10:24:00.553997",
- "modified_by": "Administrator",
- "name": "Add A Few Customers",
- "owner": "Administrator",
- "ref_doctype": "Customer",
- "slide_desc": "",
- "slide_fields": [
-  {
-   "align": "",
-   "fieldname": "customer_name",
-   "fieldtype": "Data",
-   "label": "Customer Name",
-   "placeholder": "",
-   "reqd": 1
-  },
-  {
-   "align": "",
-   "fieldtype": "Column Break",
-   "reqd": 0
-  },
-  {
-   "align": "",
-   "fieldname": "customer_email",
-   "fieldtype": "Data",
-   "label": "Email ID",
-   "reqd": 1
-  }
- ],
- "slide_title": "Add A Few Customers",
- "slide_type": "Create",
- "submit_method": ""
-}
