--- frappe/desk/setup_wizard_slide/add_a_few_suppliers/add_a_few_suppliers.json
@@ -1,61 +1,0 @@
-{
- "add_more_button": 1,
- "app": "ERPNext",
- "creation": "2019-11-15 14:45:32.626641",
- "docstatus": 0,
- "doctype": "Setup Wizard Slide",
- "domains": [
-  {
-   "domain": "Manufacturing"
-  },
-  {
-   "domain": "Services"
-  },
-  {
-   "domain": "Retail"
-  },
-  {
-   "domain": "Distribution"
-  }
- ],
- "help_links": [
-  {
-   "label": "Supplier",
-   "video_id": "zsrrVDk6VBs"
-  }
- ],
- "idx": 0,
- "image_src": "/assets/erpnext/images/illustrations/supplier.png",
- "max_count": 3,
- "modified": "2019-11-26 10:23:54.632950",
- "modified_by": "Administrator",
- "name": "Add A Few Suppliers",
- "owner": "Administrator",
- "ref_doctype": "Supplier",
- "slide_desc": "",
- "slide_fields": [
-  {
-   "align": "",
-   "fieldname": "supplier_name",
-   "fieldtype": "Data",
-   "label": "Supplier Name",
-   "placeholder": "",
-   "reqd": 1
-  },
-  {
-   "align": "",
-   "fieldtype": "Column Break",
-   "reqd": 0
-  },
-  {
-   "align": "",
-   "fieldname": "supplier_email",
-   "fieldtype": "Data",
-   "label": "Supplier Email",
-   "reqd": 1
-  }
- ],
- "slide_title": "Add A Few Suppliers",
- "slide_type": "Create",
- "submit_method": ""
-}
