--- frappe/desk/setup_wizard_slide/welcome_to_erpnext!/welcome_to_erpnext!.json
@@ -1,36 +1,0 @@
-{
- "add_more_button": 0,
- "app": "ERPNext",
- "creation": "2019-11-13 16:00:04.735347",
- "docstatus": 0,
- "doctype": "Setup Wizard Slide",
- "domains": [
-  {
-   "domain": "Manufacturing"
-  },
-  {
-   "domain": "Retail"
-  },
-  {
-   "domain": "Services"
-  },
-  {
-   "domain": "Distribution"
-  },
-  {
-   "domain": "Education"
-  }
- ],
- "help_links": [],
- "idx": 0,
- "image_src": "/assets/erpnext/images/illustrations/onboard.png",
- "max_count": 0,
- "modified": "2019-11-25 15:27:46.314226",
- "modified_by": "Administrator",
- "name": "Welcome to ERPNext!",
- "owner": "Administrator",
- "slide_desc": "Setting up an ERP can be overwhelming. But don't worry, we have got your back!<br>\nLet's setup your company.\nThis wizard will help you onboard to ERPNext in a short time!",
- "slide_fields": [],
- "slide_title": "Welcome to ERPNext!",
- "slide_type": "Information"
-}
