+++ frappe/model/sync.py
@@ -44,9 +44,7 @@ def sync_for(app_name, force=0, sync_everything = False, verbose=False, reset_pe
 			("data_migration", "data_migration_plan"),
 			("desk", "setup_wizard_slide_field"),
 			("desk", "setup_wizard_help_link"),
-			("desk", "setup_wizard_slide"),
-			("desk", "setup_wizard_slide_order"),
-			("desk", "setup_wizard_settings")):
+			("desk", "setup_wizard_slide")):
 			files.append(os.path.join(frappe.get_app_path("frappe"), d[0],
 				"doctype", d[1], d[1] + ".json"))
 
@@ -75,7 +73,7 @@ def get_doc_files(files, start_path, force=0, sync_everything = False, verbose=F
 	# load in sequence - warning for devs
 	document_types = ['doctype', 'page', 'report', 'dashboard_chart_source', 'print_format',
 		'website_theme', 'web_form', 'notification', 'print_style',
-		 'data_migration_mapping', 'data_migration_plan', 'setup_wizard_slide', 'setup_wizard_settings']
+		 'data_migration_mapping', 'data_migration_plan', 'setup_wizard_slide']
 	for doctype in document_types:
 		doctype_path = os.path.join(start_path, doctype)
 		if os.path.exists(doctype_path):
