+++ frappe/printing/setup_wizard_slide/company_letter_head/company_letter_head.json
@@ -4,20 +4,7 @@
  "creation": "2019-11-22 13:25:42.892593",
  "docstatus": 0,
  "doctype": "Setup Wizard Slide",
- "domains": [
-  {
-   "domain": "Services"
-  },
-  {
-   "domain": "Manufacturing"
-  },
-  {
-   "domain": "Retail"
-  },
-  {
-   "domain": "Distribution"
-  }
- ],
+ "domains": [],
  "help_links": [
   {
    "label": "Know more about printing and branding through letterhead",
@@ -27,7 +14,7 @@
  "idx": 0,
  "image_src": "/assets/erpnext/images/illustrations/letterhead.png",
  "max_count": 0,
- "modified": "2019-11-26 14:47:22.595626",
+ "modified": "2019-11-26 18:21:01.509703",
  "modified_by": "Administrator",
  "name": "Company Letter Head",
  "owner": "Administrator",
@@ -43,6 +30,7 @@
    "reqd": 0
   }
  ],
+ "slide_order": 20,
  "slide_title": "Company Letter Head",
  "slide_type": "Create",
  "submit_method": ""
